refactor: extract hotkey dialog component
