+++ frontend/src/hotkeys.css
@@ -22,3 +22,11 @@
 #hotkey-help button {
   margin-top: 1rem;
 }
+
+#hotkey-help kbd {
+  background: #f5f5f5;
+  border: 1px solid #ccc;
+  border-radius: 3px;
+  padding: 0 0.25rem;
+  font-family: monospace;
+}
