+++ frontend/src/index.html
@@ -3,6 +3,7 @@
 <head>
   <meta charset="UTF-8" />
   <title>Multicode Editor</title>
+  <!-- Styles for the hotkey help dialog -->
   <link rel="stylesheet" href="./hotkeys.css" />
   <style>
     body { margin: 0; font-family: sans-serif; }
